Reworked bloom pass to be closer to what's described in https://learnopengl.com/Guest-Articles/2022/Phys.-Based-Bloom. This mainly involved removing the prefilter pass, since everything is always blooming but the effect is more evident on brighter objects because they emit way more photons than other objects. Also added a filter radius as described in the website, but it introduced some weird banding artifacts so it's disabled for now (e.g. CONFIG::BloomFilterRadius is 0.0f)

+++ src/pipelines/bloom_downscaling_pipeline.cpp
@@ -42,7 +42,7 @@ namespace TANG
 		// Enable Karis average
 		VkPushConstantRange pushConstant{};
 		pushConstant.offset = 0;
-		pushConstant.size = sizeof(float);
+		pushConstant.size = sizeof(uint32_t);
 		pushConstant.stageFlags = VK_SHADER_STAGE_COMPUTE_BIT;
 
 		VkPipelineLayoutCreateInfo pipelineLayoutInfo = PopulatePipelineLayoutCreateInfo(setLayoutArray.data(), static_cast<uint32_t>(setLayoutArray.size()), &pushConstant, 1);
